+++ test/basic.test.js
@@ -67,28 +67,6 @@ function levelTest (name, level) {
     instance[name]('hello %d', 42)
   })
 
-  test('passing error at level ' + name, function (t) {
-    t.plan(2)
-    var err = new Error('myerror')
-    var instance = pino(sink(function (chunk, enc, cb) {
-      t.ok(new Date(chunk.time) <= new Date(), 'time is greater than Date.now()')
-      delete chunk.time
-      t.deepEqual(chunk, {
-        pid: pid,
-        hostname: hostname,
-        level: level,
-        type: 'Error',
-        msg: err.message,
-        stack: err.stack,
-        v: 1
-      })
-      cb()
-    }))
-
-    instance.level = name
-    instance[name](err)
-  })
-
   test('passing error with a serializer at level ' + name, function (t) {
     t.plan(2)
     var err = new Error('myerror')
